requirements.txt
absl-py==0.12.0
astor==0.8.1
click==7.1.2
Flask==1.1.1
Flask-Cors==3.0.10
gast==0.4.0
google-pasta==0.2.0
grpcio==1.36.1
gunicorn==20.0.4
h5py==2.10.0
importlib-metadata==3.7.3
Jinja2==2.11.3
json5==0.9.5
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
libsm6==1.2.1
Markdown==3.3.4
MarkupSafe==1.1.1
numpy>=1.3.0
pillow==7.0.0
protobuf==3.15.6
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
scipy==1.5.4
six==1.15.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
typing-extensions==3.7.4.3
Werkzeug==1.0.1
wrapt==1.12.1
